Update cloneVolume to use V7 of SF API
The older cloneVolume calls to the API created true clones
of the parent volume, meaning that everything was copied
including attributes.  This meant that a clone process in
Cinder was two steps:
1. Create the clone
2. Use modifyVolume to go back and update attributes, QoS etc

This isn't the most efficient way to do things any more, and it
also presents a potential issue.  If the cloneVolume encounters
an error where the Voume is created we end up never doing the
update, and now have a voume with incorrect attribute info on
the system.

This creates problems like duplicate volumes (based on UUID attribute)
for subsequent retry calls that might be issued.  We could use a
finally block and another try/catch, but newer versions of the
SolidFire API allow us to do include updated parameters on the Clone
cmd itself.  So this change just rolls that all into the single
Clone call and avoids the pitfall altogether.

We also add a check for volume on Clone failure.  There are some cases
where we might get an exception back from the API here but the clone
is actually created and exists, so before we give up, just add a
simple querie to see if it's there or not.