Add new attach/detach API calls
This patch aims to make the attach/detach process a bit
simpler and less prone to race conditions and solve
some confusion in it's usage.

For the first step we aren't implementing much, just
some new calls for attach and detach.  We'll call them
create/remove_attachment for now, and they'll just consolidate
all of the calls we used to make.

For example, instead of:
 1. reserve
 2. initialize
 3. attach

We'll just call create_attachment(volume, connector,
                                  instance_uuid, mountpoint)

For now that will simply call the existing initialize_connection
and attach methods, but the advantage here is it's a single call
for the consumer, and more importantly we can require all of the
various info we need up front.  This solves things like
multi-attach problems, because now we have all the info from
initialize, when we make the call to attach.

Eventually the old steps should be reimplemented and made more
efficient than they are currently, but if we get past this first
step we can have a stable, easy to use API to work off of that
shouldn't have to change AND best of all we don't break any
compatability.

Change-Id: I90d54257862e1815a03ca8750040f7e4e6f35c0b