Backport clone_image lock and retry SliceNotReady
This backports some bug fixes to Kilo, specifically:

https://bugs.launchpad.net/cinder/+bug/1605027
https://bugs.launchpad.net/cinder/+bug/1623618

This patch will apply a lock to the clone_image method in the
SolidFire driver.  Clone image is only used if the enable_caching
config option is set for create-volume-from-image.

One downside to the addition of the lock is... well, its' a lock.
As a result we may take a hit in performance in cases where we
used to allow multiple clones of the same image-template
simultaneoulsy, this will make those operations sequential.

This same locking method can be used on the general clone method
if needed, but it would be wise to gather some performance data
specific to the customers use-case prior to adding this to
production.