Add BlockBox
BlockBox is Cinder services running in Containers. BlockBox is a
basic template for setting that up and running it via compose.

Note there are many ways to go about doing this, and this is in no
way intended to be viewed as better than another, but is merely a
simple baseline to allow you to quickly and easily deploy a running
standalone Cinder setup.