Build LVM from loci-cinder/lvm
The Dockerfile needed to build c-vol/lvm  is now merged in the
openstack/loci-cinder project.  This change calls that repo
directly and removes the custom Dockerfile that was in the
BlockBox directory.

One thing that's challenging is that currently Docker doesn't
provide a mechanism to specify a base-image (FROM) variable.
This means that if you supply a different name/tag to the Makefile,
when it tries to build the lvm image it won't recognize the image on
your system and will instead try to pull from dockerhub.  The dockerhub
images aren't being built automatically (yet) and this won't really
work.  So for now, we'll make the naming/tagging static to keep
everything compatible.

The upcoming release of Docker will allow FROM to be overriden via
build-args, at which time we can reenable the ability to set custom
names if desired.

Change-Id: Icef6124c38c9300cd057c04bc9517e76c76070e4