Enable ability to failback SolidFire replication
This change fixes up a few things in the SolidFire driver to enable
auotomated `failback` functionality in Cinders host replication.

Previously the use case for Cinder replication was DR scenario where
the original/src cluster was "destroyed" and was now a puddle of
melted solder and plastic.  Turns out people want to test things
and like to see things go flippy floppy back and forth.  So...
this patch removes the assumption that the original cluster has
become a smoking hole in the ground.

Now, when a failover-host is called, we failover all replicated
volumes just as before, but we also demote the source (assuming
it still exists) to be a replication target.  So, any new writes
to the failed over volume will by synced back to the original.

If after some time the cloud operator would like to go back to
using the original primary, he/she can issue another failover-host
and specify the additional backend_id parameter as `default`:
    `cinder failover-host <service-host-name> --backend_id 'default'`

This will notify the driver that we want to go back to the originally
configured SF cluster and promote it back to the primary, and set our
replication targets back up again.

NOTE:  We do NOT and will NOT support individual volume failover.

********** WIP *******************
  * Could you some code simplificiation/cleanup
  * Test a few corner cases
  * Fix/Add unit tests
  * Add docs